+++ src/main/java/com/cobber/fta/plugins/LogicalTypePhoneNumber.java
@@ -87,7 +87,7 @@ public class LogicalTypePhoneNumber extends LogicalTypeInfinite  {
 		}
 
 		@Override
-		public String isValidSet(String dataStreamName, long matchCount, long realSamples, StringFacts stringFacts, Map<String, Integer> cardinality, Map<String, Integer> outliers) {
+		public String isValidSet(String dataStreamName, long matchCount, long realSamples, StringFacts stringFacts, Map<String, Long> cardinality, Map<String, Long> outliers) {
 			return (double)matchCount/realSamples >= getThreshold()/100.0 ? null : REGEXP;
 		}
 }
